fix: resolve gcc warnings flagged across the ROS buildfarm (Humble/Jazzy)
Addresses every warning category surfaced by the last Humble (Ubuntu
Jammy, gcc-11) and Jazzy (Ubuntu Noble, gcc-13) ROS buildfarm dev jobs:
-Wold-style-cast, -Wconversion, -Wsign-conversion, -Wfloat-conversion,
-Wunused-parameter, -Wunused-result, -Wdeprecated-declarations,
-Wdangling-else, -Wdangling-reference, -Wnon-virtual-dtor,
-Woverloaded-virtual, -Wsign-compare, -Wstringop-overread, -Wshadow,
-Wreturn-local-addr and -Wrange-loop-construct, spanning mrpt_bayes,
mrpt_containers, mrpt_core, mrpt_graphs, mrpt_graphslam, mrpt_hwdrivers,
mrpt_img, mrpt_maps, mrpt_math, mrpt_nav, mrpt_obs, mrpt_poses,
mrpt_serialization, mrpt_slam, mrpt_system and mrpt_viz.

Real bugs fixed along the way (not just silenced casts):

* CPose3DQuat::operator[](i) const returned a temporary `double` by
  value while its `const_iterator` bound it to a `const double&`,
  making every dereference of a const iterator dangling
  (-Wreturn-local-addr).
* CDirectedTree::Visitor and every one of its subclasses had virtual
  functions but a non-virtual destructor.
* CSerializable's schema-archive overloads of serializeTo/serializeFrom
  were silently hidden in any class using DEFINE_SERIALIZABLE, so they
  were unreachable without explicit qualification; added the missing
  `using` declarations.
* mrpt::Clock::time_point had no discoverable (ADL-reachable)
  operator<<, which made gtest's EXPECT_EQ/ASSERT_EQ fail to compile
  wherever it compared two timestamps on gcc-11/older libstdc++ (the
  actual Humble build error, not just a warning) -- this was the one
  case where mrpt::system::operator<<(TTimeStamp) already existed but
  lived in a namespace ADL never finds from outside mrpt::system; it is
  now `mrpt::operator<<`, with `mrpt::system` bringing it back in via a
  `using` declaration for existing qualified callers.
* Several TKFMethod-shaped unscoped enums (TKFMethod,
  TLaserSimulUncertaintyMethod, TMatrixTextFileFormat,
  TICPCovarianceMethod) had no fixed underlying type, so a test
  casting an intentionally-invalid value into them triggered
  "conversion is unspecified" rather than the well-defined reinterpret
  a fixed underlying type gives.

Everything else is either a widening/narrowing cast made explicit, an
unused parameter renamed to a comment, `else`-ambiguous one-liners
given braces, or a documented false-positive (e.g. a wrapper around a
cached-static-map accessor gcc's -Wdangling-reference still flags)
resolved by dropping the unnecessary named reference instead of
suppressing the warning.

Verified with an incremental colcon build of all 16 touched packages
under mrpt_common's warning set (-Wall -Wextra -Wshadow -Wtype-limits
-Wcast-align -Wparentheses -Wunused -Wpedantic -Wconversion
-Wsign-conversion -Wdouble-promotion -Woverloaded-virtual
-Wold-style-cast -Wnon-virtual-dtor): zero warnings left other than a
gcc-13/-O3 -Wmaybe-uninitialized false positive on small fixed-size
CMatrixDynamic construction that isn't present in either buildfarm log
and isn't touched by this change. All existing unit tests
(2000+ cases across the touched packages) still pass.

+++ modules/mrpt_math/include/mrpt/math/distributions.h
@@ -119,7 +119,7 @@ double KLD_Gaussians(
   const MATRIXLIKE2 cov1_inv = cov1.inverse_LLt();
   const VECTORLIKE1 mu_difs = mu0 - mu1;
   return 0.5 * (log(cov1.det() / cov0.det()) + (cov1_inv * cov0).trace() +
-                multiply_HtCH_scalar(mu_difs, cov1_inv) - N);
+                multiply_HtCH_scalar(mu_difs, cov1_inv) - static_cast<double>(N));
   MRPT_END
 }
 
@@ -226,7 +226,7 @@ void confidenceIntervals(
   out_mean = mean(data);
   const auto x_min = data.minCoeff();
   const auto x_max = data.maxCoeff();
-  const auto binWidth = (x_max - x_min) / histogramNumBins;
+  const auto binWidth = (x_max - x_min) / static_cast<double>(histogramNumBins);
 
   const std::vector<double> hitsNormalized =
       mrpt::math::histogram(data, x_min, x_max, histogramNumBins);
@@ -240,8 +240,8 @@ void confidenceIntervals(
   ASSERT_(it_high != Hc.end());
   const size_t idx_low = std::distance(Hc.begin(), it_low);
   const size_t idx_high = std::distance(Hc.begin(), it_high);
-  out_lower_conf_interval = x_min + idx_low * binWidth;
-  out_upper_conf_interval = x_min + idx_high * binWidth;
+  out_lower_conf_interval = x_min + static_cast<double>(idx_low) * binWidth;
+  out_upper_conf_interval = x_min + static_cast<double>(idx_high) * binWidth;
 
   MRPT_END
 }
@@ -265,7 +265,7 @@ void confidenceIntervalsFromHistogram(
 
   const auto x_min = *histogramCoords.begin();
   const auto x_max = *histogramCoords.rbegin();
-  const auto binWidth = (x_max - x_min) / histogramCoords.size();
+  const auto binWidth = (x_max - x_min) / static_cast<double>(histogramCoords.size());
 
   std::vector<double> Hc;
   cumsum(histogramNormalizedHits, Hc);  // CDF
@@ -277,8 +277,8 @@ void confidenceIntervalsFromHistogram(
   ASSERT_(it_high != Hc.end());
   const size_t idx_low = static_cast<size_t>(std::distance(Hc.begin(), it_low));
   const size_t idx_high = static_cast<size_t>(std::distance(Hc.begin(), it_high));
-  out_lower_conf_interval = x_min + idx_low * binWidth;
-  out_upper_conf_interval = x_min + idx_high * binWidth;
+  out_lower_conf_interval = x_min + static_cast<double>(idx_low) * binWidth;
+  out_upper_conf_interval = x_min + static_cast<double>(idx_high) * binWidth;
 
   MRPT_END
 }
